pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "PyFlink Stream Processing Workshop"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=2.2.0",
    "psycopg-binary>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.2.2",
    "sodapy>=2.2.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pgcli>=4.4.0",
]
